disable all startup services (#37)
It doesn't make sense for  any of the daemons in charmed-mongodb snap to start up automatically in the snap. They require specific configurations and an up and a running mongos instance. We can keep them enabled, but they will show an error state:
```
ubuntu@ip-172-31-37-213:~$ sudo snap logs charmed-mongodb.pbm-agent
2023-09-16T14:25:39Z systemd[1]: Started Service for snap application charmed-mongodb.pbm-agent.
2023-09-16T14:25:39Z charmed-mongodb.pbm-agent[4119638]: 2023/09/16 14:25:39 Error: Parse command line parameters: required flag --mongodb-uri not provided
2023-09-16T14:25:39Z systemd[1]: snap.charmed-mongodb.pbm-agent.service: Deactivated successfully.
```